add hostname to auto pxe script
when booting osie we need to know the hostname to update the
metadata and hw-info on self-test.

today we don't have any way to pass on a single hardware information
that would be passed to the osie when first boot.

this extention will add the id={hostname} to the /proc/cmdline which
will allow hwinfo process to extract this field and update selftest.

on v1beta1 we will have a more generic way to pass in kernel args on
the hardware level.

we used to use facility but for some-reason this is not passed
on by https://github.com/tinkerbell/dhcp which is the current
kube-backend for tink:

https://github.com/tinkerbell/dhcp/blame/d8def67df9bdce1ff5269cd475094f08f25900bf/backend/kube/kube.go#L276